+++ NEWS.md
@@ -1,6 +1,13 @@
 profvis (development version)
 =============================
 
+* `print()` gains a `sort` argument. Use `print(profvis(f()), sort =
+  "alphabetical")` to sort alphabetically within each row of the
+  flamegraph instead of by time. Sorting alphabetically favours box
+  merging and makes it easier to see the big picture (#115). Set the
+  `profvis.sort` global option to `"aphabetical"` to change the
+  default sort.
+
 profvis 0.3.7
 =============